[FIX] cleanup enum options typing
