+++ frontend/src/components/Attribute/AttributeEntityForm.vue
@@ -58,7 +58,7 @@
   <AttributeEnumOptionsInput
     v-if="data.data_type === 'ENUM'"
     :ref="(el: InputRef) => (refs.enumOptions = el)"
-    v-model="enumOptions"
+    v-model="data.enum_options"
     :attribute-id="'id' in attribute ? attribute.id : undefined"
   />
   <EntityInput
@@ -87,7 +87,7 @@
 
 <script setup lang="ts">
 import { useI18n } from 'src/composables/useI18n';
-import { watch, computed, ref } from 'vue';
+import { watch, ref } from 'vue';
 import { extend } from 'quasar';
 import EntityInput from '../Entity/Edit/EntityInput.vue';
 import EntityToggle from '../Entity/Edit/EntityToggle.vue';
@@ -103,9 +103,18 @@ import AttributeValidationRuleInput from './AttributeValidationRuleInput.vue';
 import AttributeLegendInput from './AttributeLegendInput.vue';
 import AttributeDefaultValueInput from './AttributeDefaultValueInput.vue';
 import AttributeEnumOptionsInput from './AttributeEnumOptionsInput.vue';
-import type { EnumOptionInput } from './AttributeEnumOptionsInput.vue';
+import type { EnumOptionInput } from './enumOption';
+import type { DistributiveOmit } from 'src/utils/typescriptUtils';
 
 type Attribute = AttributeModalEditProps['attribute'];
+
+// While editing, enum options can be unsaved (no `id`) and carry a transient `_uid`,
+// so the form's working copy uses `EnumOptionInput` instead of the stricter server
+// shape on `Attribute`. The parent reconciles this when transforming the emitted
+// data into mutation variables.
+type AttributeFormData = DistributiveOmit<Attribute, 'enum_options'> & {
+  enum_options: EnumOptionInput[];
+};
 
 export interface AttributeEntityFormProps {
   attribute: Attribute;
@@ -113,7 +122,7 @@ export interface AttributeEntityFormProps {
 
 const props = defineProps<AttributeEntityFormProps>();
 const emits = defineEmits<{
-  change: [data: Attribute];
+  change: [data: AttributeFormData];
 }>();
 // for defineExpose() see below
 
@@ -127,16 +136,9 @@ const initialData = {
   legend: props.attribute.legend,
   default_value: props.attribute.default_value,
   enum_options: props.attribute.enum_options,
-} as Attribute;
+} as AttributeFormData;
 
-const data = ref<Attribute>(extend(true, {}, initialData));
-
-const enumOptions = computed<EnumOptionInput[]>({
-  get: () => data.value.enum_options,
-  set: (val) => {
-    data.value.enum_options = val as typeof data.value.enum_options;
-  },
-});
+const data = ref<AttributeFormData>(extend(true, {}, initialData));
 
 const refs = ref<{ [key: string]: InputRef | null }>({
   name: null,
